+++ lib/rage-rb.rb
@@ -28,7 +28,7 @@ module Rage
   end
 
   def self.env
-    @__env ||= Rage::Env.new(ENV["RAGE_ENV"] || ENV["RAILS_ENV"] || ENV["RACK_ENV"] || "development")
+    @__env ||= Rage::Env.new(ENV["RAGE_ENV"])
   end
 
   def self.groups
@@ -45,8 +45,30 @@ module Rage
 
   def self.load_middlewares(rack_builder)
     config.middleware.middlewares.each do |middleware, args, block|
+      # in Rails compatibility mode we first check if the middleware is a part of the Rails middleware stack;
+      # if it is - it is expected to be built using `ActionDispatch::MiddlewareStack::Middleware#build`, but Rack
+      # expects the middleware to respond to `#new`, so we wrap the middleware into a helper module
+      if Rage.config.internal.rails_mode
+        rails_middleware = Rails.application.config.middleware.middlewares.find { |m| m.name == middleware.name }
+        if rails_middleware
+          wrapper = Module.new do
+            extend self
+            attr_accessor :middleware
+            def new(app, *, &)
+              middleware.build(app)
+            end
+          end
+          wrapper.middleware = rails_middleware
+          middleware = wrapper
+        end
+      end
+
       rack_builder.use(middleware, *args, &block)
     end
+  end
+
+  def self.code_loader
+    @code_loader ||= Rage::CodeLoader.new
   end
 
   module Router