+++ lib/rage/all.rb
@@ -6,10 +6,11 @@ require_relative "fiber"
 require_relative "fiber_scheduler"
 require_relative "configuration"
 require_relative "request"
+require_relative "response"
 require_relative "uploaded_file"
 require_relative "errors"
 require_relative "params_parser"
-require_relative "fiber_wrapper"
+require_relative "code_loader"
 
 require_relative "router/strategies/host"
 require_relative "router/backend"
@@ -21,7 +22,12 @@ require_relative "router/node"
 require_relative "controller/api"
 
 require_relative "logger/text_formatter"
+require_relative "logger/json_formatter"
 require_relative "logger/logger"
+
+require_relative "middleware/fiber_wrapper"
+require_relative "middleware/cors"
+require_relative "middleware/reloader"
 
 if defined?(Sidekiq)
   require_relative "sidekiq_session"