+++ lib/rage/logger/text_formatter.rb
@@ -8,15 +8,20 @@ class Rage::TextFormatter
 
   def call(severity, timestamp, _, message)
     logger = Thread.current[:rage_logger]
-    tags = logger[:tags]
+    tags, context = logger[:tags], logger[:context]
+
+    if !context.empty?
+      context_msg = ""
+      context.each { |k, v| context_msg << "#{k}=#{v} " }
+    end
 
     if final = logger[:final]
       params, env = final[:params], final[:env]
       if params
-        return "[#{tags[0]}] timestamp=#{timestamp} pid=#{@pid} level=info method=#{env["REQUEST_METHOD"]} path=#{env["PATH_INFO"]} controller=#{params[:controller]} action=#{params[:action]} status=#{final[:response][0]} duration=#{final[:duration]}\n"
+        return "[#{tags[0]}] timestamp=#{timestamp} pid=#{@pid} level=info method=#{env["REQUEST_METHOD"]} path=#{env["PATH_INFO"]} controller=#{params[:controller]} action=#{params[:action]} #{context_msg}status=#{final[:response][0]} duration=#{final[:duration]}\n"
       else
         # no controller/action keys are written if there are no params
-        return "[#{tags[0]}] timestamp=#{timestamp} pid=#{@pid} level=info method=#{env["REQUEST_METHOD"]} path=#{env["PATH_INFO"]} status=#{final[:response][0]} duration=#{final[:duration]}\n"
+        return "[#{tags[0]}] timestamp=#{timestamp} pid=#{@pid} level=info method=#{env["REQUEST_METHOD"]} path=#{env["PATH_INFO"]} #{context_msg}status=#{final[:response][0]} duration=#{final[:duration]}\n"
       end
     end
 
@@ -32,13 +37,6 @@ class Rage::TextFormatter
         i += 1
       end
       tags_msg << " timestamp=#{timestamp} pid=#{@pid} level=#{severity}"
-    end
-
-    context = logger[:context]
-
-    if !context.empty?
-      context_msg = ""
-      context.each { |k, v| context_msg << "#{k}=#{v} " }
     end
 
     "#{tags_msg} #{context_msg}message=#{message}\n"