+++ lib/rage/middleware/fiber_wrapper.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 ##
-# The middleware wraps every request in a Fiber and implements the custom defer protocol with Iodine.
-# Scheduling fibers in a middleware allows the framework to be compatibe with custom Rack middlewares.
+# The middleware wraps every request in a separate Fiber. It should always be on the top of the middleware stack,
+# as it implements a custom defer protocol, which may break middlewares located above.
 #
 class Rage::FiberWrapper
   def initialize(app)
@@ -16,9 +16,11 @@ class Rage::FiberWrapper
     fiber = Fiber.schedule do
       @app.call(env)
     ensure
+      # notify Iodine the request can now be resumed
       Iodine.publish(Fiber.current.__get_id, "", Iodine::PubSub::PROCESS) if Fiber.current.__yielded?
     end
 
+    # the fiber encountered blocking IO and yielded; instruct Iodine to pause the request
     if fiber.alive?
       [:__http_defer__, fiber]
     else