+++ lib/rage/router/dsl.rb
@@ -7,6 +7,8 @@ class Rage::Router::DSL
 
   def draw(&block)
     Handler.new(@router).instance_eval(&block)
+    # propagate route definitions to Rails for `rails routes` to work
+    Rails.application.routes.draw(&block) if Rage.config.internal.rails_mode
   end
 
   ##
@@ -283,7 +285,7 @@ class Rage::Router::DSL
       end
 
       _module, _path, _only, _except, _param = opts.values_at(:module, :path, :only, :except, :param)
-      raise ":param option can't contain colons" if _param&.include?(":")
+      raise ":param option can't contain colons" if _param.to_s.include?(":")
 
       _only = Array(_only) if _only
       _except = Array(_except) if _except