+++ lib/rage/setup.rb
@@ -2,22 +2,10 @@ Iodine.patch_rack
 
 require_relative "#{Rage.root}/config/environments/#{Rage.env}"
 
+# Run application initializers
+Dir["#{Rage.root}/config/initializers/**/*.rb"].each { |initializer| load(initializer) }
 
-# load application files
-app, bad = Dir["#{Rage.root}/app/**/*.rb"], []
-
-loop do
-  path = app.shift
-  break if path.nil?
-
-  require_relative path
-
-# push the file to the end of the list in case it depends on another file that has not yet been required;
-# re-raise if only errored out files are left
-rescue NameError
-  raise if (app - bad).empty?
-  app << path
-  bad << path
-end
+# Load application classes
+Rage.code_loader.setup
 
 require_relative "#{Rage.root}/config/routes"